+++ built_value_generator/lib/src/serializer_source_class.dart
@@ -8,7 +8,6 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value_generator/src/serializer_source_field.dart';
-import 'package:meta/meta.dart';
 
 part 'serializer_source_class.g.dart';